i28: extract jounal read to separate module
