+++ src/journal.rs
@@ -1,0 +1,86 @@
+use std::fs::OpenOptions;
+use crate::daily_score;
+use crate::daily_score::DailyScore;
+use std::{io, fmt};
+use std::io::{BufRead, BufReader};
+use std::error::Error;
+
+#[derive(Debug)]
+pub enum JournalError {
+    CannotOpenFile { file_path: String, open_error: io::Error },
+    CannotReadLine { file_path: String, read_error: io::Error },
+    DailyScoreParseError { line: String, daily_score_parse_error: daily_score::ParseError },
+}
+
+impl fmt::Display for JournalError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            Self::CannotOpenFile { file_path, open_error: _ } => write!(f, "cannot open journal file '{}'", file_path),
+            Self::CannotReadLine { file_path, read_error: _ } => write!(f, "cannot read line from journal file '{}'", file_path),
+            Self::DailyScoreParseError { line, daily_score_parse_error: _ } => write!(f, "cannot parse daily score data '{}'", line),
+        }
+    }
+}
+
+impl std::error::Error for JournalError {
+    fn source(&self) -> Option<&(dyn Error + 'static)> {
+        match self {
+            Self::CannotOpenFile { file_path: _, open_error } => Some(open_error),
+            Self::CannotReadLine { file_path: _, read_error } => Some(read_error),
+            Self::DailyScoreParseError { line: _, daily_score_parse_error } => Some(daily_score_parse_error),
+        }
+    }
+}
+
+pub fn read(file_path: &str) -> Result<Vec<DailyScore>, JournalError> {
+    let mut records = Vec::<DailyScore>::new();
+
+    let file = OpenOptions::new()
+        .read(true)
+        .open(file_path)
+        .map_err(|open_error| JournalError::CannotOpenFile {
+            file_path: file_path.to_string(),
+            open_error,
+        })?;
+
+    let reader = BufReader::new(file);
+    for line in reader.lines() {
+        let line_string = line.map_err(|read_error| JournalError::CannotReadLine {
+            file_path: file_path.to_string(),
+            read_error,
+        })?;
+
+        let daily_score =
+            DailyScore::parse(line_string.as_str())
+            .map_err(|daily_score_parse_error|
+                     JournalError::DailyScoreParseError {
+                         line: line_string.clone(),
+                         daily_score_parse_error,
+                     })?;
+
+        records.push(daily_score);
+    }
+    Ok(records)
+}
+
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn errors_display() {
+        let file_path = String::from("path/to/file");
+        let io_error = io::Error::new(io::ErrorKind::Other, "error text");
+        let another_io_error = io::Error::new(io::ErrorKind::Other, "error text");
+        let line = String::from("foo bar baz");
+        let daily_score_parse_error = daily_score::ParseError::MissingDateTime;
+
+        assert_eq!(JournalError::CannotOpenFile { file_path: file_path.clone(), open_error: io_error }.to_string(),
+            "cannot open journal file 'path/to/file'");
+        assert_eq!(JournalError::CannotReadLine { file_path, read_error: another_io_error }.to_string(),
+            "cannot read line from journal file 'path/to/file'");
+        assert_eq!(JournalError::DailyScoreParseError { line, daily_score_parse_error }.to_string(),
+            "cannot parse daily score data 'foo bar baz'");
+    }
+}
