+++ src/mood_command.rs
@@ -1,13 +1,15 @@
-use std::{io, fmt};
-use std::io::{BufRead, BufReader};
-use std::fs::OpenOptions;
+//use std::{io, fmt};
+//use std::io::{BufRead, BufReader};
+//use std::fs::OpenOptions;
 use std::error::Error;
 use std::collections::HashSet;
+use std::fmt;
 
-use crate::daily_score;
-use crate::daily_score::DailyScore;
+//use crate::daily_score;
+//use crate::daily_score::DailyScore;
 use crate::mood_report::MoodReport;
 use crate::Config;
+use crate::journal;
 
 mod plot;
 
@@ -36,62 +38,31 @@ impl MoodReportType {
 
 #[derive(Debug)]
 pub enum MoodCommandError {
-    CannotOpenFile { file_path: String, open_error: io::Error },
-    CannotReadLine { file_path: String, read_error: io::Error },
-    DailyScoreParseError { line: String, daily_score_parse_error: daily_score::ParseError },
+    JournalReadError(journal::JournalError),
 }
 
 impl std::error::Error for MoodCommandError {
     fn source(&self) -> Option<&(dyn Error + 'static)> {
         match self {
-            Self::CannotOpenFile { file_path: _, open_error } => Some(open_error),
-            Self::CannotReadLine { file_path: _, read_error } => Some(read_error),
-            Self::DailyScoreParseError { line: _, daily_score_parse_error } => Some(daily_score_parse_error),
-        }
+            Self::JournalReadError(journal_error) => Some(journal_error),
+         }
     }
 }
 
 impl fmt::Display for MoodCommandError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
-            Self::CannotOpenFile { file_path, open_error: _ } => write!(f, "cannot open journal file '{}'", file_path),
-            Self::CannotReadLine { file_path, read_error: _ } => write!(f, "cannot read line from journal file '{}'", file_path),
-            Self::DailyScoreParseError { line, daily_score_parse_error: _ } => write!(f, "cannot parse daily score data '{}'", line),
+            Self::JournalReadError(_journal_error) => write!(f, "cannot parse journal"),
         }
     }
 }
 
 impl MoodCommand {
     pub fn run(self) -> Result<(), MoodCommandError> {
-        let mut records = Vec::<DailyScore>::new();
+        let daily_scores = journal::read(&self.config.file_path)
+            .map_err(|journal_error| MoodCommandError::JournalReadError(journal_error))?;
 
-        let file = OpenOptions::new()
-            .read(true)
-            .open(self.config.file_path.as_str())
-            .map_err(|open_error| MoodCommandError::CannotOpenFile {
-                file_path: self.config.file_path.clone(),
-                open_error,
-            })?;
-
-        let reader = BufReader::new(file);
-        for line in reader.lines() {
-            let line_string = line.map_err(|read_error| MoodCommandError::CannotReadLine {
-                file_path: self.config.file_path.clone(),
-                read_error,
-            })?;
-
-            let daily_score =
-                DailyScore::parse(line_string.as_str())
-                .map_err(|daily_score_parse_error|
-                         MoodCommandError::DailyScoreParseError {
-                             line: line_string.clone(),
-                             daily_score_parse_error,
-                         })?;
-
-            records.push(daily_score);
-        }
-
-        let mood_report = MoodReport { daily_scores: &records, tags: &self.tags };
+        let mood_report = MoodReport { daily_scores: &daily_scores, tags: &self.tags };
 
         let (caption, data) = match self.report_type {
             MoodReportType::Monthly => ("30-days mood:", mood_report.thirty_days_mood()),
@@ -112,26 +83,5 @@ impl MoodCommand {
         }
 
         Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn errors_display() {
-        let file_path = String::from("path/to/file");
-        let io_error = io::Error::new(io::ErrorKind::Other, "error text");
-        let another_io_error = io::Error::new(io::ErrorKind::Other, "error text");
-        let line = String::from("foo bar baz");
-        let daily_score_parse_error = daily_score::ParseError::MissingDateTime;
-
-        assert_eq!(MoodCommandError::CannotOpenFile { file_path: file_path.clone(), open_error: io_error }.to_string(),
-            "cannot open journal file 'path/to/file'");
-        assert_eq!(MoodCommandError::CannotReadLine { file_path, read_error: another_io_error }.to_string(),
-            "cannot read line from journal file 'path/to/file'");
-        assert_eq!(MoodCommandError::DailyScoreParseError { line, daily_score_parse_error }.to_string(),
-            "cannot parse daily score data 'foo bar baz'");
     }
 }
